Entrega_3/requirements.txt
python-dotenv
psycopg2-binary
pandas
requests
# y otras dependencias necesarias

